added activities to project to implement the game; the code still contains errors
