+++ app/src/main/AndroidManifest.xml
@@ -8,15 +8,16 @@
         android:label="@string/app_name"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".MainActivity">
+        <activity android:name="csci4020.shawnbickel.assignment1.MenuActivity">
             <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-
-                <category android:name="android.intent.category.LAUNCHER" />
+                <action android:name="android.intent.action.MAIN"/>
+                <category android:name="android.intent.category.LAUNCHER"/>
             </intent-filter>
         </activity>
-        <activity android:name=".BlackJackDetails" />
-        <activity android:name=".BlackjackGame"></activity>
+        <activity android:name="csci4020.shawnbickel.assignment1.BlackJackDetails" />
+        <activity android:name="csci4020.shawnbickel.assignment1.GameImplementation">
+
+        </activity>
     </application>
 
 </manifest>
