+++ app/src/main/java/csci4020/shawnbickel/assignment1/BlackJackDetails.java
@@ -1,4 +1,4 @@
-package csci4020.shawnbickel.assignment1.blackjack;
+package csci4020.shawnbickel.assignment1;
 
 import android.content.Intent;
 import android.net.Uri;
@@ -6,6 +6,7 @@ import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.widget.Button;
+import csci4020.shawnbickel.assignment1.blackjack.R;
 
 public class BlackJackDetails extends AppCompatActivity {
