--- app/src/main/java/csci4020/shawnbickel/assignment1/blackjack/BlackjackGame.java
@@ -1,13 +1,0 @@
-package csci4020.shawnbickel.assignment1.blackjack;
-
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-
-public class BlackjackGame extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_blackjack_game);
-    }
-}
